services/ocr-api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-api"
version = "0.1.0"
description = "FastAPI service for OCR intake, normalization and validation"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "numpy>=1.26,<3.0",
  "opencv-python-headless>=4.10,<5.0",
  "pymupdf>=1.24,<2.0",
  "openai>=1.54,<2.0",
  "pillow>=10.4,<12.0",
  "pillow-heif>=0.18,<1.0",
  "pypdf>=5.0,<6.0",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
  "python-multipart>=0.0.9,<1.0",
  "rapidocr-onnxruntime>=1.2.3,<1.3",
  "uvicorn[standard]>=0.30,<1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0"
]
rapidocr = [
  "rapidocr-onnxruntime>=1.2.3,<1.3"
]
paddleocr = [
  "paddleocr>=2.8,<3.0"
]
doctr = [
  "python-doctr>=0.9,<1.0"
]
azure = [
  "azure-ai-documentintelligence>=1.0.0b3,<2.0",
  "azure-core>=1.30,<2.0"
]
google = [
  "google-cloud-documentai>=2.30,<3.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
